move useEffect channel redirect to ServerList
